Grafana UI: DashboardFileUpload => FileUpload
