+++ packages/grafana-ui/src/components/index.ts
@@ -152,6 +152,7 @@ export { Switch } from './Switch/Switch';
 export { Checkbox } from './Forms/Checkbox';
 
 export { TextArea } from './TextArea/TextArea';
+export { FileUpload } from './FileUpload/FileUpload';
 
 // Legacy forms
